fixes several mypy errors
